Extract Chunk into separate file
