+++ test_runner/src/main/kotlin/ftl/args/CalculateShardsResult.kt
@@ -1,3 +1,5 @@
 package ftl.args
 
+import ftl.shard.Chunk
+
 data class CalculateShardsResult(val shardChunks: List<Chunk>, val ignoredTestCases: IgnoredTestCases)
