+++ test_runner/src/main/kotlin/ftl/shard/Chunk.kt
@@ -1,0 +1,6 @@
+package ftl.shard
+
+data class Chunk(val testMethods: List<TestMethod>) {
+    val testStringList = testMethods.map { it.name }
+    val size get() = testMethods.size
+}
