+++ test_runner/src/test/kotlin/ftl/run/GenericTestRunnerTest.kt
@@ -1,6 +1,6 @@
 package ftl.run
 
-import ftl.args.Chunk
+import ftl.shard.Chunk
 import ftl.args.IArgs
 import ftl.run.platform.common.beforeRunMessage
 import ftl.shard.TestMethod
